[SCRUM-190] Rename : 패키지 이름 변경
