+++ src/main/java/com/kkinikong/be/community/repository/elasticsearch/CommunityPostSearchRepository.java
@@ -3,7 +3,7 @@ package com.kkinikong.be.community.repository.elasticsearch;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 import org.springframework.stereotype.Repository;
 
-import com.kkinikong.be.community.domain.elasticsearch.CommunityPostDocument;
+import com.kkinikong.be.community.domain.document.CommunityPostDocument;
 
 @Repository
 public interface CommunityPostSearchRepository
